requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
google-generativeai
python-dotenv